+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - \#3242 - Treat "value" attribute in server-side validation errors as general
   error
 - \#3304 - Marathon UI hangs after being open for a while
+- \#3233 - Cannot remove a constraint
 
 ## 0.15.6 - 2016-02-23
 - \#3192 - Adapt default Mem/CPU settings